back_end_test/requirements.txt
pandas==2.3.1
openpyxl==3.1.5
langchain_google_genai==2.1.9
langchain_experimental==0.3.4
tabulate==0.9.0
python-dotenv==1.1.1
langchain==0.3.27
langgraph==0.6.5
langchain-text-splitters==0.3.9
langchain-core==0.3.74
langchain_chroma==0.2.5